requirements.txt
# Base 
fastapi==0.120.3
uvicorn==0.38.0
pydantic==2.12.3

# for PostgreSQL
asyncpg==0.30.0

# Authentication and Security
passlib==1.7.4
python-jose==3.5.0 
bcrypt==3.2.0 
PyJWT==2.10.1
cryptography==46.0.3 

# for file handling
python-multipart==0.0.20
pillow==12.0.0

# Configuration
python-dotenv==1.2.1

email-validator==2.3.0